feat: extract knowledge table component
